Fix VDiff2 DeleteByUUID Query
It was incorrect and thus deleting every row in the vdiff table.
IIRC, this was a mistake on my part as I was first using a WHERE
based query and then had to move to LEFT JOIN usage as there may
not always be vdiff_table records. When doing so, however, I left
the last AND clause in place when it should have become the sole
predicate in a new WHERE clause after the LEFT JOIN condition.

Signed-off-by: Matt Lord <mattalord@gmail.com>