mysqlctl: Move more to use built in MySQL client
executeMysqlScript was still using a `mysql` client binary for
initialization, but there's no reason it needs to do that. We can also
move this to our built in MySQL client to be faster and have less
dependencies here.

Follow up to #13178 where we already moved things for just the schema
commands.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>