Remove additional unused Dockerfile entries
The `docker/k8s` directory also seems to trace back to when Helm was
still a thing. It also doesn't actually have a `vtorc` container which
indicates it's not actually used since `vtorc` is actually required
these days.

From the looks of it, the same applies to `docker/mini` and that is
unused as well.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>