icuregex: Fix invalid slice creation
Parse the structure so we can create buffers with the proper size and
never with infinite sizes. While this was not the immediate cause of the
race error, it's better to create with the right slice size also for
debugging when digging into it.

The real fix here is that the size of `algorithmicRange` includes the
size of the struct itself, so if we want to get the remaining slice size
it needs to subtract this value.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>