backup: Rework version detection
It's not guaranteed that MySQL will always be up when doing a restore,
this is shown in some of the failing tests with the changes so far.

This means we need a way to get the MySQL version without having it
running. We can though rely in `mysqlctld` being up. So this adds a new
RPC to retrieve the version number through invoking the binary. This
ensures we still get the right version also if we're running from the
tablet as we'd execute the RPC.

We also have less of a versioning issue here, as the vttablet + mysql
are always upgraded together and the new mysqlctl is injected into the
mysql container if needed so we don't have a version mismatch. So it's
safe to do this in one step.

Also decided to make the upgrade safe an explicit flag for `vtbackup` as
well since overloading the existing flag was wonky and not really what
we'd want. This makes it more explicit and safe.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>