planbuilder: Pass through type information as well
With the type information we can better decide if we need to print the
collation change or not, since it's only needed for textual columns.

This also fixes a number of cases where we used to use the weight_string
but we no longer need to since we now pass through the correct type
information in more cases.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>