Fix potential panics due to "Fail in goroutine after test completed"
The root issue here is that some of the functions we call before the `cancel(); wg.Wait()` section include calls to assertions from the `require` subpackage, which, unlike their `assert` counterparts, end up calling `t.FailNow()` which aborts the test run via `runtime.Goexit()`
If this happens while another goroutine is still running, the test "completes", and if that other goroutine has any additional assertions that fail, this panic occurs.

Luckily, `runtime.GoExit` still executes the deferred function calls in the exiting goroutine, so as long as we defer the context cancelation and wait, we will ensure that that other goroutine exits before the main goroutine of the test case.

Since these tests sometimes need to run code both before and after the "cancel and wait for goroutine" section, we need to doubly-nest a function definition to ensure we have a defer and it executes when we want it to.

Signed-off-by: Andrew Mason <andrew@planetscale.com>