Add dry-run/monitoring-only mode for TxThrottler
This change adds the ability to run the transaction throttler in dry-run/
monitoring-only mode. This means that it will only emit metrics related to
whether throttling would or would not have take place, but will not actually
throttle. This can be useful when deploying, to get a sense of how often
throttling can happen without actually having it happen.

Signed-off-by: Eduardo J. Ortega U <5791035+ejortegau@users.noreply.github.com>