Ensure to call servenv.Init when needed
Since https://github.com/vitessio/vitess/pull/12987 profiling is not
loaded implicitly through an import but only explicitly. This means we
need to make sure that `servenv.Init` is called.

It's not called for all binaries which is an oversight. It also explains
why this wasn't found with https://github.com/vitessio/vitess/pull/12987
since that was tested with binaries like `vttablet` and `vtgate` and it
was assumed that this loading was done everywhere properly, which it
sadly was not.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>