evalengine: Fix JSON weight string computation
The external WEIGHT_STRING function runs weight strings on JSON as if
they are on the JSON string representation. This matches that behavior
just for the evalengine function specifically.

We also still want to have a correct weight string implementation that
does honor the MySQL JSON type ordering rules. So we also test here some
more cases and explicit type conversions for those functions. This
uncovered a small missing but for temporal types which we didn't
convert properly before but these conversion tests now expose.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>