mysql: Refactor dependencies
This starts splitting up the large `mysql` package into various
subpackages that can handle more specific cases. The first is to extract
replication concerns and sqlerror.

These two packages now deal with parsing of GTIDs, positions and
replication status and the latter with all the MySQL level error types.

With this split, `vtorc` can actually build without loading in all the
collations information as it doesn't actually need any of that. Before
it was pulled in because it would need `go/mysql` for replication and
error information.

We can still further split off some main concerns, such as a
`go/mysql/flavor` package and a `go/mysql/conn` package to split off
specific bits for flavor detection and the actual MySQL connection.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>