evalengine: Refactor out int conversion functions
The `ToInt64` and `ToUint64` conversion functions end up loading the
whole evalengine in many places, even though only the conversion is
needed.

Instead we move this as a function to `sqltypes.Value` itself. The
casting logic to best effort convert a type to those is the same as
depending on the full `evalengine` logic to do it. It's much simpler
though and removes the dependency from many places.

Additionally this moves a few more bits to the sidecar constants with
some utility functions. This allows us to decouple `mysqlctl` fully from
the `evalengine` as well.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>