sqlparser: Track if original default value is a literal
We want to keep if the user explicitly specified an expression or
something we treat as a literal, so we can serialize it back to the
exact same format.

This is because for example `now()` and `(now())` do have a semantic
difference. The latter works as a default for `datetime(6)` with a
precision, while the former does not and would need `now(6)` in the
literal form.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>