docker/bootstrap: remove --no-cache flag
It seems like this was added back in 2015 when vitess used to use raw
go get commands to build the image, the reasoning behind it was to ensure
that the image has the most recent dependencies. This is no longer needed
because in the image we correctly add go.* dep manifests and call go mod download.

PR where this was added:
https://github.com/vitessio/vitess/pull/665

Signed-off-by: Amir Abushareb <yields@icloud.com>