icuregex: Update to ICU 73
This updates the ICU data to ICU 73. The most significant change here
is that in ICU 70 the emoji properties were split off into a separate
file, so we need to include that file here too and add a loader for it.

Additionally the regex code was reviewed from 69 to 73 to see if there's
any significant differences. There were a few typo fixes in comments and
some crash assertions were changed to errors instead.

MySQL also updated to bundling ICU 73 with the 8.0.34 release.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>