Ensure hexval and int don't share bindvar
There was some existing code that attempted to disambiguate literal
values as map keys by prefixing a single quote. Since this proved
to not be correctly handling all situations, we thought it appropriate
to key by the Literal itself. This ensures that a literal type is also
considered first class in disambiguation.

We chose to use the Literal rather than *Literal because literals with
the same type and value are intended to share a BindVar.

There's now slightly more data in the key since it is a struct with
a Type field of Int, but since the key previously was an arbitrary length
string, the difference should be negligble. Running the benchmarks
didn't seem to indicate anything of concern.

Signed-off-by: William Martin <williammartin@github.com>