servenv: Remove double close() logic
The servenv.Run function already runs the close hooks once it finishes.
This means there's no use for deferring a `.Close()` since it double
executes all close hooks.

In fact, it leads to panics. For example, vttablet always panics on each
shutdown because the connection to the topo is already closed.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>