evalengine: Improve the typing situation for functions
There are a number of functions like `unix_timestamp` that have more
subtle typing properties and who need additional information. This adds
logic to track sizing so we can return the correct type also for those
functions.

This removes the last ambiguous typing we have here and fixes a number
of queries.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>