evalengine: Fix nullable handling
A number of expressions and function implementations in the evalengine
were not correctly propagating the nullable checks. For example the
multiplication operation didn't do this. This can lead to crashes in the
evalengine.

Here the flags have been audited and set properly. New test inputs are
also added to ensure we exercise more of these paths.

Signed-off-by: Manan Gupta <manan@planetscale.com>
Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>