Allow for passing in the MySQL shutdown timeout
We have a hardcoded timeout today of 300 seconds when we call mysqladmin
during a MySQL shutdown. This means that for example when we have fast
shutdown disabled for an upgrade, we still kill after 5 minutes and
don't allow for a longer shutdown.

Large tablets with a lot of data can definitely take longer than 5
minutes, so this hardcoded timeout leads to failures in this case which
means we can't easily upgrade the tablet and need to do a full
replacement from a backup.

By allowing this to be configured, we can make sure that we can give it
longer timeouts when we want to do an upgrade.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>