mysqlctl: Error out on stale socket
When we see a stale socket file, it's not useful to try and start MySQL
since we're going to fail to start it anyway.

In that case it's better to error out immediately and either a human or
computer operator can try to fix the situation and retry instead of
waiting for `wait_time` which is useless at this point.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>