build: Allow compilation on Windows
Vitess is not always just used standalone, but also sometimes for
components. For example the parser, schemadiff, the MySQL connection parser etc. are
components used separately.

Even though these parts have no OS specific bits (like process
management etc. for tablet management and MySQL), we can't compile this
on Windows due to dependencies on things like syslog in places.

These changes allow for building Vitess on Windows. This does not in any
way indicate actual support for Windows for Vitess or even plans to ever
do that in the future, it's solely here to make it easier to use Vitess
parts as a library.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>