mysql: Refactor out usage of servenv
The servenv package is something that is use for server environments
like vtgate, vttablet etc. But the go/mysql package should really be
independent code for things like the MySQL protocol bits.

This change refactors things so that go/mysql doesn't depend anymore on
servenv. This makes it easier to be used as a library for example.

The remaining bits here are in collations, which is something to tackle
separately as it's very invasive.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>