Remove all non-test usage of collations.Local()
This removes all non-test usage of collations.Local() in favor of
explicitly injecting it everywhere. This may look noisy, but that's
because it is and it now makes the dependencies we already have as
global explicit instead.

It shows the entanglement in the codebase which is in a way good,
because it shows what we should look at if we need to clean it up or not
to decouple things better.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>