Remove collations.Local()
Now that we inject the collation environment in all non test
environments, we can remove collations.Local() which depends on the
global version number config. This removes the servenv dependency from
collations.

As alternative, we now have collations.MySQL8() which returns a static
MySQL 8 based collations environment that can be used for testing
instead.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>