Cleanup of dead code
Go recently announced https://go.dev/blog/deadcode and this does a first
pass and cleans up a bunch of things.

Not everything it identifies is really dead code for us. We have a bunch
of test implementations of interfaces for example, which is seen as dead
code when running without the `-test` flag.

But with the `-test` flag, it doesn't find code that only is used from
tests.

What it points at is a good starting point though for what to then
manually verify and clean up. That's what I did here.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>