sqlparser: Refactor out servenv and inject everywhere
This is a refactor similar as was done with collations, to remove the
servenv dependency from sqlparser and to make sqlparser usable entirely
with no global state for other dependencies.

This means we now create a sqlparser instance with the appropriate
settings instead of modifying global configuration flags for things like
the version and for truncation settings.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>