evalengine: Fix week overflow
Some of the last days in the year can fall in the first week of the next
year. If that happens, there are various modes when we should return 53
as the week number and not week 1 for the next year.

This can always be 53. There are weeks which already contain 53 weeks,
but it can never be 54 in those cases as in years with 53 weeks, there's
never a last day of the year that falls in the first week of next year.
It then always falls in week 53 already.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>