Fix a number of CodeQL warnings
This addresses a number of CodeQL warnings. We have a concurrency field
which is using inconsistent types across different fields, which means
we end up casting and triggering warnings around overflows. We can use
the smallest type here, namely int32 since that is for sure sufficient
for real concurrency values. It avoids warnings like:

https://github.com/vitessio/vitess/security/code-scanning/267

Also fixes the parsing bug identified in
https://github.com/vitessio/vitess/security/code-scanning/266 where we
parse with `ParseUint` instead of `ParseInt`.

Last one is
https://github.com/vitessio/vitess/security/code-scanning/268 which is
fixed by making sure we parse it as the type MySQL sends down and to
keep using that consistently.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>