Improve err extraction logic
The previous regexp had superfluous wildcard matching at the beginning
and end. This results in eagerly matching the whole string and
subsequently matches the last occurance of the match.

Instead we don't want to use the wildcard so the default left to right
matching is used and we get the first occurence.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>