schemadiff: Introduce environment
We already had the parser that needed to be injected, but in order to
properly handle MySQL 5.7 we also need to inject collations
configuration. Instead of adding more arguments, we introduce the
schemadiff environment struct and use that to inject the things like the
current parser and current collation environment and default collation.

Also need to refactor the rest that ends up using schemadiff to inject
it similarly.

The core goal here is that the sidecar can properly inject the right
collation as well so we can unblock fixing the case sensitivity there.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>