Fix failover test for reparenting
This test is pretty broken, as it uses MariaDB replication, but we're
running this with MySQL.

The reason the test doesn't outright fail is a combination of two
things. First, the test doesn't assert that the command to change
replication don't error, so it returns an error.

Secondly, this is also affect by the multi execute issue of #14948, so
it's not even guaranteed the queries run properly.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>