vtenv: Introduce vtenv for passing in collation & parser information
We have many places taking already 3 arguments for the collation env,
parser & MySQL version. This is due to the refactors to pass in explicit
state.

This introduces `vtenv` to refactor things a bit and slightly clean it
up. We now have one thing that can carry all three dependencies and also
potential future additional ones where we need it.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>