Fix parsing at negative edge case
We didn't parse the case properly where we had mod 2 base and the value
was math.MinInt64, we'd reject that incorrectly. We need to check
different for if we have a minus or not.

Also adds automated tests for all these edge cases by testing the 1000
numbers before & after the limits for each type for each possible base.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>