mysql/datetime: Improve TIME parsing logic
In places inside the `evalengine` we need to know if a `TIME` instance
was properly parsed, if a value could be extracted from input (but it
wasn't entirely valid) or if there's enirely invalid input.

Before we only had one flag that indicated the strict parsing. So we
couldn't really discern partial parsing success from entire failure. We
tried that with checking for a zero time, but that's wrong. A zero time
is still an entirely valid time and can be the result of successful
partial parsing.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>