vtexplain: Fix setting up the column information
In https://github.com/vitessio/vitess/pull/13312 we fixed how we encode
names in information_schema queries. The problem there though is that
the `GetColumnNamesQuery` string was also used inside `vtexplain`. It's
usage there was not updated for the new escaping logic.

This in turn leads to broken queries during `vtexplain` which then leads
to a lot of logs. The issue here is not that we log a lot, we log that
we actually do have a problem here which is the case.

This fixes the underlying issue of not correctly generating the query as
we should.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>