srvtopo: Setup metrics in init() functions
We currently create a single `srvTopoServer` inside `NewTabletServer` to
avoid panics on duplicate metrics.

This is wrong though, since the topo server created is tied to the
context given, and if that is cancelled, it will be stopped. We will
then never start a new one.

The reason this was set up like that is because of metrics. So instead
of what we did before, we need to create metrics once. All other metrics
are created in `init()` functions and we need to do the same here.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>