Set `VReplicationStreamState` to Copying on resumption
Previously, the `VReplicationStreamState` would initialize as Copying,
but if stopped and restarted, the state shows as Running (despite the
workflow status showing as Copying). This mismatch can cause errors in
reporting and monitoring (depending on which status one recognizes).

This change sets the `VReplicationWorkflowState` (and thereby
`VReplicationStreamState`) as Copying when the copy workflow is resumed.
The workflow status and `VReplicationStreamState` stat are now synced.

Signed-off-by: Tyler Coleman <tyler@planetscale.com>