Setup individual timeouts for each binary
Until now we've had the same timeouts for shutdown etc. across the
different binaries. This doesn't provide the necessary flexibility
though, since for example `mysqlctld` will need longer timeouts.

This is because it needs to shut down MySQL and we need it to honor the
MySQL shutdown timeout as well. This means this can take significantly
longer to ensure the shutdown is clean.

It's necessary to do this, to ensure things like being able to upgrade
MySQL since that depends on clean shutdowns.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>