vtorc: Clean up more unused code
This goes a bit further than previous just `deadcode` cleanup, but it
digs deeper into which parts are actually unused. We don't seem to use
any of the binary log file positioning logic, since we all use GTID
based logic so that can all be cleaned up.

This then removes a bunch of other unused fields etc. as well.

Lastly, we rename the data_center attribute to cell since that's the
internal Vitess terminology.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>