Fix additional type size handling for decimals
This audits more cases in the evalengine where we create datetime
related or decimal types and ensures we initialize them with the correct
sizing.

This is necessary to ensure that the type information is correct further
downstream also for the evalengine compiler since there are cases where
we have to type switch based in input sizes.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>