Fix GRPC deprecations
This change means "opening" a connection no longer really already dials
to the actual other side, so we should see if this has any other
significant effects if we instead dial on first request.

The old stuff is deprecated though, so we have to bite the bullet at
some point. Alternatively, we essentially copy this initial connect
logic into our side instead.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>
Signed-off-by: Florent Poinsard <florent.poinsard@outlook.fr>