Fix the race condition during vttablet startup
This avoids the problem where the connection pool is poisoned when we
check for the MySQL port, by avoiding to use the pool in the first
place. We only ever run this once at startup, so we can create a new
connection here and then dispose of it once we've retrieved the port.

That way we know the connection pool is still clean and doesn't have any
problems.

Fixes #15730

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>