evalengine: Add support for enum and set
The evalengine currently doesn't handle enum and set types properly.
The comparison function always returns 0 at the moment and we don't
consider ordering of elements etc.

Here we add two new native types to the evalengine for set and enum and
instantiate those appropriately. We also ensure we can compare them
correctly.

In case we don't have the schema information with the values, we do a
best effort case of depending on the string representation. This is not
correct always of course, but at least makes equality comparison work
for those cases and only ordering is off in that scenario.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>