Revert GRPC context changes
These changes were made assuming we could use `grpc.NewClient`
everywhere. But we had failing test for that. In a specific case, we
replace it with `grpc.Dial` but that's not the same.

If we can't move it all to `grpc.NewClient` yet, we have to keep the
context setup we do as well, or otherwise `grpc.Dial` could hang
indefinitely even if the caller context is cancelled since it gets
started with the background context.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>