grpc: Always pass through context for dialer
The change in https://github.com/vitessio/vitess/pull/15780 made me
audit the rest of the places where we dial with GRPC today where we
don't pass in a context yet. This is something we should do, since the
Dial can block otherwise indefinitely even on context cancel in
the caller.

This fixes those cases and even cleans up a TODO that mentioned this
needed to be done in a specific spot.

Once we move to grpc.NewClient in a more piecemeal fashion, we can
probably clean up / undo a lot of these cases but until then we should
do this properly with a context passed in.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>