Switch to use semisync source / replica plugins
The plugins with the old terminology are deprecated and will be removed
in the future. In MySQL 8.4.0 also the old terminology for regular
replication is removed.

This starts the move to use the new style. Since MySQL 8.0.26 the new
semisync plugin is available, so we should start using it from that
version on.

This means we also need new default MySQL cnf files to set up things
correctly. We add here a mysql8026.cnf and also a pre-emptive
mysql84.cnf. The latter is needed to start the work in the future for
MySQL 8.4.0. The main change here is that the deprecated
mysql_native_password plugin needs to be explicitly enabled.

Removing our usage of mysql_native_password is another separate
significant effort. The main issue there is how we want to deal with
certificates etc. which end up being required for replication if you
want to use caching_sha2_password which adds a whole layer of
complexity we've never had to deal with.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>