projection: Return correct collation information
When we have a project, we correctly convert the value to the connection
collation using:

```go
c.Value(vcursor.ConnCollation())
```

But we then don't update / change the collation on the fields
information. This needs to match the actual data, otherwise the client
might corrupt the data.

This ensures we add the same rules for the resulting collation.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>