mysqlctld: Remove unneeded resets in init_db.sql
This removes the resets for the startup script that we use. These were
long ago added for a MariaDB 10.3 workaround which is not supported as a
managed tablet anyway anymore.

It also would only needed for an existing node to join a cluster that is
otherwise manually emptied. If that's the case, the reset should also be
done manually anyway too.

Removing this also ensures we have no logic here that would start
failing in the future on MySQL 8.4.0 which removed these statements.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>