mysqlctl: Improve backup restore compatibility check
This check has to be smarter to allow certain backup restore scenarios.
Starting from MySQL 8.0.35, it's allowed to restore into older versions
within the same LTS stream. This is something we didn't allow so far,
but we need to for managing the more recent MySQL releases correctly.

This also implies that we don't need upgrade safe flags for these cases
as well and allows it even if that flag is off now.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>