go/cmd: Audit and fix context.Background() usage
This removes a bunch of the context.Background() usage for the command
line entrypoints. In general, we should use the command context (which
normally is context.Background(), but it's more semantically accurate).

There's a few cases where we need more fixes. Specifically in vtgate
where we want to setup a cancellable context and cancel it when we shut
down. This is the one that ends up running things like the topo watcher
and this ensures things are closed appropriately.

Similarly in vtcombo we apply similar fixes so that we always correctly
cancel the context on shutdown and the same for vttablet.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>