Fix schemadiff semantics handling
The problem here is that we store entries in `.Tables` with their
unescaped names, which is entirely safe to do here.

However, we would use `sqlparser.String` to retrieve them. Now that
normally works for table names, unless a table name is a SQL keyword. In
that case, it turns into the escaped version with backticks and the
whole table can't be found for semantic analysis.

This can happen for example if you name a table `Order`.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>