Handle new restrict_fk_on_non_standard_key option in MySQL 8.4
MySQL 8.4 adds restrict_fk_on_non_standard_key for which the details can
be found here:

https://dev.mysql.com/doc/refman/8.4/en/server-system-variables.html#sysvar_restrict_fk_on_non_standard_key

It prevents the use of non-unique keys or partial keys as foreign keys.

For regular MySQL replication though, it does the following:

Implication for MySQL Replication.  When a foreign key is created on a
nonstandard key on the primary because restrict_fk_on_non_standard_key
is OFF, the statement succeeds on the replica regardless of any setting
on the replica for this variable.

Given how this works fir regular replication, we should mirror this
behavior for vreplication and always allow this. So we set it up so we
can do this with similar logic as to how we disable foreign key checks
and we reset it back afterwards.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>