mysqlctl: open backup files with fadvise(2) and FADV_SEQUENTIAL
On Linux, the fadvise(2) syscall can be leveraged to hint to the kernel
to readahead disk pages since the file is intended to be read
sequentially.

This works well with our use of opening this file, then slamming it into
`io.Copy`.

In benchmarking, this alone yields around ~50% improvement on the read
side without other tunings.

In a naive test, the setup was simply an io.Copy with an os.File for
src, and io.Discard for dest:

  no fadvise: 222.41 MB/s
with fadvise: 344.69 MB/s

This is with no other tunings to the kernel.

Refs #3931

Signed-off-by: Matt Robenolt <matt@ydekproductions.com>